refactor: file upload 로직 hooks에서 관리하도록 수정
